+++ app/src/main/java/com/example/sudoku/domain/usecase/Grid.kt
@@ -2,7 +2,7 @@ package com.example.sudoku.domain.usecase
 
 import com.example.sudoku.Config
 
-data class Digit(val value: Short = 0, val fixed: Boolean = false)
+data class Digit(val value: Int = 0, val fixed: Boolean = false)
 
 typealias Grid = List<List<Digit>>
 
@@ -15,7 +15,7 @@ private fun Grid.isValidRow(rowIndex: Int, value: Int): Boolean {
     val readValues = Array(Config.GRID_LENGTH + 1) { index -> index == value }
 
     get(rowIndex).forEach { digit ->
-        digit.value.toInt().let { digitValue ->
+        digit.value.let { digitValue ->
             if (0 < digitValue && readValues[digitValue]) return false
             readValues[digitValue] = true
         }
@@ -28,7 +28,7 @@ private fun Grid.isValidColumn(columnIndex: Int, value: Int): Boolean {
     val readValues = Array(Config.GRID_LENGTH + 1) { index -> index == value }
 
     for (i in 0 until Config.GRID_LENGTH) {
-        get(i)[columnIndex].value.toInt().also { digitValue ->
+        get(i)[columnIndex].value.also { digitValue ->
             if (0 < digitValue && readValues[digitValue]) return false
             readValues[digitValue] = true
         }
@@ -47,7 +47,7 @@ private fun Grid.isValidRegion(rowIndex: Int, columnIndex: Int, value: Int): Boo
 
     for (y in (minColumnIndex until maxColumnIndex)) {
         for (x in (minRowIndex until maxRowIndex)) {
-            val digitValue = get(x)[y].value.toInt()
+            val digitValue = get(x)[y].value
             if (0 < digitValue && readValues[digitValue]) return false
             readValues[digitValue] = true
         }
